ENH: Refactor Install rule variables to only be used in install commands.
Added a second set of variables named Slicer_<blah>_DIR which contain the appropriate directory prefixes such as bin lib share. These variables can be used reference file in the build or install trees such as the input/output of a configure command. Slicer_INSTALL_<blah> should only be used for install commands as on the Mac these variables will be prefixed with Slicer.app\Contents which we likely do not want in the root of the build tree.

svn-url: http://viewvc.slicer.org/viewvc.cgi/Slicer4?view=revision&revision=16771
git-svn-id: http://svn.slicer.org/Slicer4/trunk@16771 3bd1e089-480b-0410-8dfb-8563597acbee