BUG: fix a crash caused by an overlay file with unassigned labels.
When loading some FreeSurfer scalar overlay files, the reader 
returns a warning that some labels were unassigned. This was being
treated as an error, but a valid node was being returned from the
Models logic instead of NULL, causing a crash in the IO file. 
Fixed that issue then changed the response to that warning
to return valid nodes and color tables and that seemed to 
fix the issue.

Issue #2851



svn-url: http://viewvc.slicer.org/viewvc.cgi/Slicer4?view=revision&revision=21608
git-svn-id: http://svn.slicer.org/Slicer4/trunk@21608 3bd1e089-480b-0410-8dfb-8563597acbee