BUG: Fix py_ThresholdThreadingTest
Probably as a result of VTK Python wrapper updates, None is not accepted anymore as vtkMRMLSelectionNode::SetReferenceSecondaryVolumeID input.
Since SetReferenceSecondaryVolumeID was just a workaround for fixing Python wrapping issues, there is no more reason to use it and we switch to SetSecondaryVolumeID.
Changes are performed similarly for volume and label volume.

svn-url: http://viewvc.slicer.org/viewvc.cgi/Slicer4?view=revision&revision=27432
git-svn-id: http://svn.slicer.org/Slicer4/trunk@27432 3bd1e089-480b-0410-8dfb-8563597acbee