BUG: Allow reading of FreeSurfer models saved with vtk extension
When imported into Slicer, FreeSurfer models are read using a vtkMRMLFreeSurferModelStorageNode.
When saving, the storage node has no write methods, resulting in it falling back on the parent implementation and saving as a vtk (or other vtkMRMLModelNode format).
Trying to read the saved vtk files will cause an error, since vtkMRMLFreeSurferModelStorageNode does not handle non-FreeSurfer file types.

This fixes the issue by adding a workaround to revert to the parent implementation if the file extension is vtk.

svn-url: http://viewvc.slicer.org/viewvc.cgi/Slicer4?view=revision&revision=28771
git-svn-id: http://svn.slicer.org/Slicer4/trunk@28771 3bd1e089-480b-0410-8dfb-8563597acbee