BUG: Fix FreeSurfer surface loading transformation by switching to NiBabel
NiBabel contains a more complete implementation for loading FreeSurfer models, such as being able to read in the additional information from the surface file footer.
This commit switches from using the vtkFSSurfaceReader implementation to using NiBabel for surface model loading. Other types of FreeSurfer files are currently still loaded using the existing classes.

A benefit of this change is that it is no longer needed to specify a reference volume when loading in a surface, since the information is read from the surface file footer.

Re #17